Context Warn Fixes
