+++ client/src/pages/EventosPage.jsx
@@ -5,11 +5,9 @@ import {useEventos} from '../context/EventoContext'
 function EventosPage() {
   const {eventos, loadEventos} = useEventos()
 
-  useEffect(() => {
-    console.log("EventosPage");
-    
+  useEffect(() => {    
     loadEventos();
-  }, []);
+  }, [loadEventos]);
 
   function renderMain() {
     if (eventos.length === 0) return <h1>No hay eventos aún</h1>
